+++ src/main/java/net/emilsg/clutter/item/custom/FoodWithEffectItem.java
@@ -35,6 +35,6 @@ public class FoodWithEffectItem extends Item {
 
     @Override
     public int getMaxUseTime(ItemStack stack) {
-        return (int) useTimeInTicks;
+        return useTimeInTicks;
     }
 }
